Update AN2_Cargo_Drop.sqf
Fixes an issue where DZMS supply crate marker gets spammed on map. I don't remember this being an issue so maybe I messed something up with some customizations I made, but this has fixed it for me at least.
The fix makes sense to me since before the fix; you're checking if _fallCount exists, which it's a local variable inside the if statement meaning it doesn't exist outside of that if statement, causing it to never return false.